+++ s4e-backend/src/main/java/pl/cyfronet/s4e/service/ProductCategoryService.java
@@ -7,17 +7,16 @@ import pl.cyfronet.s4e.properties.FileStorageProperties;
 @Service
 @RequiredArgsConstructor
 public class ProductCategoryService {
-    private final static String ICON_EXTENSION_WITH_DOT = ".svg";
+    public final static String ICON_EXTENSION_WITH_DOT = ".svg";
 
     private final FileStorageProperties fileStorageProperties;
 
     public String getIconPath(String iconName) {
-        return String.join(
-                "/",
+        return String.join("",
                 fileStorageProperties.getBucket(),
+                "/",
                 fileStorageProperties.getKeyPrefixProductsCategoriesIcons(),
-                iconName,
-                ICON_EXTENSION_WITH_DOT
+                iconName + ICON_EXTENSION_WITH_DOT
         );
     }
 }